requirements.txt
# Core dependencies
python-binance==1.0.19
ccxt==4.2.25
pandas==2.1.4
numpy==1.26.3
redis==5.0.1

# ML/AI
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.2.0
tensorflow==2.15.0
stable-baselines3==2.2.1
gym==0.26.2
optuna==3.5.0
shap==0.44.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
timescale-vector==0.0.1

# Monitoring & Alerts
streamlit==1.30.0
plotly==5.18.0
python-telegram-bot==20.7
schedule==1.2.1

# Web & API
aiohttp==3.9.1
websockets==12.0
fastapi==0.109.0
uvicorn==0.27.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
asyncio==3.4.3
